Refactoring AbstractLinearSGDModel and Trainer to extract AbstractSGDModel and Trainer which operate on Parameters. This will allow future non-linear additions to Tribuo's SGD models.
